_reset() and _read_bit() should release the bus sooner to avoid contention with the driving device
While it's ok to hold the bus longer its better for everyone if the bus is released
so the 1-wire device can drive the bus as soon as its ready rather than contending with
the master for a short period of time.

Change cleans up oscilloscope bus traces and removes a short duration of ~0.4V where the device
and master are driving the bus to opposite levels.